ENH: Update vtkMRMLText* nodes and add ensure backwards compatibility with Stable version of Slicer
vtkMRMLText* related classes are now only built in the stable version of slicer (< 4.11.0).
Once the Texts module has been integrated into the Slicer core, these classes can be removed fully.
The following string is used in all locations that can be removed:
"// TODO: When the stable version of Slicer includes these classes, they can be removed from SlicerOpenIGTLink"
