+++ OpenIGTLinkIF/MRML/Text/vtkMRMLTextNode.cxx
@@ -1,0 +1,169 @@
+/*=auto=========================================================================
+
+  Portions (c) Copyright 2009 Brigham and Women's Hospital (BWH) All Rights Reserved.
+
+  See Doc/copyright/copyright.txt
+  or http://www.slicer.org/copyright/copyright.txt for details.
+
+  Program:   3D Slicer
+  Module:    $RCSfile: vtkMRMLTextNode.cxx,v $
+  Date:      $Date: 2006/03/19 17:12:29 $
+  Version:   $Revision: 1.3 $
+
+=========================================================================auto=*/
+
+// SlicerOpenIGTLink MRML includes
+#include "vtkMRMLTextNode.h"
+#include "vtkMRMLTextStorageNode.h"
+
+// MRML includes
+#include <vtkMRMLScene.h>
+
+// VTK includes
+#include "vtkObjectFactory.h"
+#include "vtkXMLUtilities.h"
+
+const int MAX_STRING_LENGTH_FOR_SAVE_WITHOUT_STORAGE_NODE = 256;
+
+//----------------------------------------------------------------------------
+vtkMRMLNodeNewMacro(vtkMRMLTextNode);
+
+//-----------------------------------------------------------------------------
+vtkMRMLTextNode::vtkMRMLTextNode()
+  : Text("")
+  , Encoding(VTK_ENCODING_US_ASCII)
+  , ForceCreateStorageNode(CreateStorageNodeAuto)
+{
+}
+
+//-----------------------------------------------------------------------------
+vtkMRMLTextNode::~vtkMRMLTextNode()
+{
+  this->SetText("");
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::SetText(const std::string &text, int encoding/*-1*/)
+{
+  vtkDebugMacro( << this->GetClassName() << " (" << this << "): setting Text to " << text);
+
+  int wasModifying = this->StartModify();
+  if (encoding >= 0)
+    {
+    this->SetEncoding(encoding);
+    }
+  if (this->Text == text)
+    {
+    this->EndModify(wasModifying);
+    return;
+    }
+  this->Text = text;
+  this->InvokeCustomModifiedEvent(vtkMRMLTextNode::TextModifiedEvent);
+  this->Modified();
+  this->EndModify(wasModifying);
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::SetEncoding(int encoding)
+{
+  vtkDebugMacro(<< this->GetClassName() << " (" << this << "): setting encoding to " << encoding);
+  int clampedEncoding = std::max(VTK_ENCODING_UNKNOWN, std::min(encoding, VTK_ENCODING_UNKNOWN));
+  if (this->Encoding != clampedEncoding)
+    {
+    this->Encoding = clampedEncoding;
+    this->InvokeCustomModifiedEvent(vtkMRMLTextNode::TextModifiedEvent);
+    this->Modified();
+    }
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::ReadXMLAttributes(const char** atts)
+{
+  int disabledModify = this->StartModify();
+  Superclass::ReadXMLAttributes(atts);
+  vtkMRMLReadXMLBeginMacro(atts);
+  vtkMRMLReadXMLStdStringMacro(text, Text);
+  vtkMRMLReadXMLIntMacro(encoding, Encoding);
+  vtkMRMLReadXMLEndMacro();
+  this->EndModify(disabledModify);
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::WriteXML(ostream& of, int nIndent)
+{
+  Superclass::WriteXML(of, nIndent);
+  vtkMRMLWriteXMLBeginMacro(of);
+  if (!this->GetStorageNode())
+    {
+    vtkMRMLWriteXMLStdStringMacro(text, Text);
+    }
+  vtkMRMLWriteXMLIntMacro(encoding, Encoding);
+  vtkMRMLWriteXMLEndMacro();
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::Copy(vtkMRMLNode* anode)
+{
+  int disabledModify = this->StartModify();
+  Superclass::Copy(anode);
+  vtkMRMLCopyBeginMacro(anode);
+  vtkMRMLCopyStringMacro(Text);
+  vtkMRMLCopyIntMacro(Encoding);
+  vtkMRMLCopyEndMacro();
+  this->EndModify(disabledModify);
+}
+
+//----------------------------------------------------------------------------
+void vtkMRMLTextNode::PrintSelf(ostream& os, vtkIndent indent)
+{
+  Superclass::PrintSelf(os, indent);
+  vtkMRMLPrintBeginMacro(os, indent);
+  vtkMRMLPrintStdStringMacro(Text);
+  vtkMRMLPrintIntMacro(Encoding);
+  vtkMRMLPrintEndMacro();
+}
+
+//---------------------------------------------------------------------------
+std::string vtkMRMLTextNode::GetDefaultStorageNodeClassName(const char* vtkNotUsed(filename))
+{
+  if (!this->Scene)
+    {
+    return "";
+    }
+
+  if (this->ForceCreateStorageNode == CreateStorageNodeNever)
+    {
+    return "";
+    }
+
+  if (this->ForceCreateStorageNode == CreateStorageNodeAuto)
+    {
+    int length = this->Text.length();
+    if (length < MAX_STRING_LENGTH_FOR_SAVE_WITHOUT_STORAGE_NODE)
+      {
+      return "";
+      }
+    }
+
+  return "vtkMRMLTextStorageNode";
+}
+
+//---------------------------------------------------------------------------
+vtkMRMLStorageNode* vtkMRMLTextNode::CreateDefaultStorageNode()
+{
+  if (!this->Scene)
+    {
+    return nullptr;
+    }
+
+  if (!this->ForceCreateStorageNode)
+    {
+    int length = this->Text.length();
+    if (length < MAX_STRING_LENGTH_FOR_SAVE_WITHOUT_STORAGE_NODE)
+      {
+      return nullptr;
+      }
+    }
+  return vtkMRMLTextStorageNode::SafeDownCast(
+    this->Scene->CreateNodeByClass(this->GetDefaultStorageNodeClassName().c_str()));
+}
