+++ OpenIGTLinkRemote/Logic/CMakeLists.txt
@@ -17,6 +17,15 @@ LIST(APPEND ${KIT}_TARGET_LIBRARIES
   vtkSlicerOpenIGTLinkIFModuleLogic
   )
 
+if (Slicer_VERSION VERSION_LESS "4.11.0")
+  #TODO: When the stable version of Slicer includes the Texts module/widgets, they can be removed from SlicerOpenIGTLink
+  list(APPEND ${KIT}_INCLUDE_DIRECTORIES
+    ${vtkSlicerOpenIGTLinkIFModuleMRML_SOURCE_DIR}/Text
+	${vtkSlicerOpenIGTLinkIFModuleMRML_BINARY_DIR}/Text
+    )
+endif()
+
+
 #-----------------------------------------------------------------------------
 SlicerMacroBuildModuleLogic(
   NAME ${KIT}
